# lib/api-zod/src/generated/types/album.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Album {
  id: number;
  name: string;
  description: string;
  /** @nullable */
  coverUrl: string | null;
  sortOrder: number;
}