# lib/api-zod/src/generated/types/loveNoteInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LoveNoteInput {
  /** @minLength 1 */
  title: string;
  content: string;
  sortOrder?: number;
}