# lib/api-zod/src/generated/types/mediaUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MediaUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  date?: string;
  /** @nullable */
  albumId?: number | null;
  isPublic?: boolean;
}