# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  content: string;
}