# lib/api-zod/src/generated/types/musicInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MusicInput {
  /** @minLength 1 */
  title: string;
  artist: string;
  url: string;
}